refactor: injection-zone
